+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "start": "node ./bin/www"
   },
+  "type": "module",
   "dependencies": {
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
