+++ src/app.js
@@ -1,0 +1,36 @@
+import createError from 'http-errors';
+import express, { json, urlencoded, static as static_ } from 'express';
+import { join } from 'path';
+import cookieParser from 'cookie-parser';
+import logger from 'morgan';
+
+import readApiRouter from './api/read.js';
+import createApiRouter from './api/create.js';
+
+const app = express();
+
+app.use(logger('dev'));
+app.use(json());
+app.use(urlencoded({ extended: false }));
+app.use(cookieParser());
+
+app.use('/api', readApiRouter)
+app.use('/api', createApiRouter)
+
+// catch 404 and forward to error handler
+app.use(function(req, res, next) {
+  next(createError(404));
+});
+
+// error handler
+app.use(function(err, req, res, next) {
+  // set locals, only providing error in development
+  res.locals.message = err.message;
+  res.locals.error = req.app.get('env') === 'development' ? err : {};
+
+  // render the error page
+  res.status(err.status || 500);
+  res.render('error');
+});
+
+export default app;
